Fix error when unbinding quoted keybindings
Keybindings that contain characters that need to be escaped, are quoted
in the output of `tmux list-keys`. For example, `M-#` and `M-$` are
listed as `"M-#"` and `"M-$"`. Moreover, quotes in the keybindings are
escaped, e.g. `M-"` and `"` are listed as `"M-\""` and `\"`.

`tmux unbind-key` will not accept these (as they are) and will report
something like

```
unknown key: "M-#"
```

We therefore need to strip the surrounding quotes and the escaping
backslashes before giving it to `tmux unbind-key`. This solves GitHub
issue #1.