Fix error when unbinding keybindings with escaped characters
Some characters are escaped when listing with `tmux list-keys`, e.g.
`M-"`, `#` and `$` are outputted as `"M-\""`, `\#` and `\$`,
respectively. Backslashes needs to be removed before running `tmux
unbind-key`, otherwise the following error may appear

```
unknown key: M-\\
```

Semicolons are escaped and they need to be that for `tmux unbind-key`.
Therefore we treat them specially (and keep the backslash). This fixes
GitHub issue #1.