Don't set the left status bar multiple times
If the main script is loaded multiple times (e.g. manually), we would
set the left status bar multiple times as well. This would result in
multiple icons in the status bar.

We can prevent this by checking for our status string in the current
status, and only set it once when it's absent. This solves GitHub
issue #4.